# Conflicts:
#	src/main/java/com/warehousesystem/app/controller/WarehouseController.java
#	src/main/java/com/warehousesystem/app/service/impl/WarehouseGoodServiceImpl.java
#	src/main/resources/application-local.yaml
#	src/main/resources/application.yaml

+++ src/main/java/com/warehousesystem/app/dto/WarehouseGoodCreateDto.java
@@ -34,5 +34,5 @@ public class WarehouseGoodCreateDto {
 
     @NotNull
     @Min(value = 1, message = "Quantity cannot be negative or zero")
-    private Long quantity;
+    private BigDecimal quantity;
 }
