+++ src/main/java/com/warehousesystem/app/dto/WarehouseGoodFullDto.java
@@ -40,7 +40,7 @@ public class WarehouseGoodFullDto {
 
     @NotNull
     @Min(value = 1, message = "Quantity cannot be negative or zero")
-    private Long quantity;
+    private BigDecimal quantity;
 
     @NotNull
     private CurrencyType currency;
@@ -50,6 +50,4 @@ public class WarehouseGoodFullDto {
 
     @JsonFormat(pattern = "dd.MM.yyyy HH:mm:ss")
     private LocalDateTime creationTime;
-
-
 }
